pyproject.toml
[project]
name = "conway2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alive-progress>=3.3.0",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
]
